backend/requirements-test.txt
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-httpx==0.34.0
